Add structure for graphs and nodes
