+++ PL2/PL2-shared/src/main/java/nl/tudelft/pl2016gr2/model/GraphInterface.java
@@ -1,0 +1,12 @@
+package nl.tudelft.pl2016gr2.model;
+
+public interface GraphInterface {
+	
+	public AbstractNode getNode(int id);
+	
+	public int getSize();
+	
+	public void addNode(AbstractNode node);
+	
+	public AbstractNode getRoot();
+}
